NN-1862 Fixing query parameter name.
This is effectively a breaking change as the api went deployed to prod earlier this week. It's unlikey to be used.
I could potentially keep the old param as an alias but it seems a bit overkill.

In the future, I'll probably use the ApiIgnore or Api(hidden=true) annotations to hide whilst in development