NN-1910 Added an endpoint to retrieve adjudication details
This endpoint currently has an ApiIgnore annotation to allow tweaks if the frontend requires it.

The cucumber test is quite light and only checks a few elementary aspects of the returned details.
Instead of testing every element in the cucumber test, I have added a marshalling test to ensure that serializing the model results in the correct json structure.

As we already have a test ensuring the model is populated correctly from the DB this seems appropriate.