DCS-188 Reduce warnings around large caches
The cache configuration was set to 20,000 objects for each cache and that number of references is shared between all items that are present in that cache.
A couple of the caches allowed up to 10,000 complex objects, which meant this threshold was easily surppassed.

We could have increased the threshold but there are performance warnings around this - there’s a comment that say's it takes roughly 1 second to process 1mb of
 objects in LRUMemoryStore and it creates a lock on the memory store for that time period (in prod the cache size can be roughly 30-50mb)

So instead, after talking to Mike, we’ve reduced the cache sizes down to 1000 items

I've also removed a couple of unused caches.