GDPR-41 Adding POST endpoint to request population of pending deletion queue.
This cannot be a GET request since the process to generate the ids for
offenders pending deletion is a fairly long running process. We have
chosen to POST the request with an immediate 202 Accepted response
and the resulting ids will be published onto a topic which the
Data Compliance Service can listen for.