WIP DT-635 Audit updates from all users
All users with scope `proxy-user` will continue to have write access and
only Nomis users will connect to the DB as the user (and fully control
the audit columns on an update via the set_context stuff).

For non-Nomis users, the user id will still be stamped on the created by
field but will also appear in the AUDIT_CLIENT_USER_ID.

If any updates require a Nomis user in order to be allowed, these are
currently rejected with a 404 if the user is not a Nomis user.  I don't
see any reason to change this or beef it up with some kind of annotation
guarding these updates.  If I try to update with a Delius user and get a
`404 User not found` then I think that's self-explanatory.