GDPR-57 Adding ability to search offender record's free text fields. (#583)
As part of the GDPR work, we need to check free text fields for certain
trigger words related to moratoria. This uses Oracle's REGEXP_LIKE
function to search free text fields for a provided regex. It only needs
to search one offender's record at a time so I don't need anything too
heavyweight, but I have considered options such as Apache Lucene (pull
all of the free text out for an offender, index in-memory and search).