SDI-378 Set movement sequence explicitly
It seems that previously movement time was trusted more than movement sequence in determining the order of movemenbts. Having found some data that contradicts this I've decided to work off movement sequence instead. This makes more sense to me as sequence is generated by the database but time is created by humans.