WIP: proposed patch to behaviour from Andy Marke
This is the change to behaviour proposed by Andy Marke so that a service
connecting using `client_credentials` with the `VIEW_ADJUDICATIONS` role
will be able to view adjudications via the

`https://api-preprod.prison.service.justice.gov.uk/api/offenders/{NOMS}/adjudications`

endpoint.

The tests describing this change in behaviour are yet to be added.