MAP-494 Refactor capacity check methods in AgencyInternalLocation (#2018)
* MAP-494 Refactor capacity check methods in AgencyInternalLocation

This commit removes the boolean parameter 'treatZeroOperationalCapacityAsNull' from the capacity check methods in AgencyInternalLocation class. Now the methods consider operational capacity as zero when it is null. The related tests and class methods that use those methods have also been updated accordingly.

* MAP-494 Remove test for zero operational capacity

The test case for zero operational capacity in AgencyInternalLocationTest.java has been removed. This test was found unnecessary as zero operational capacity is not a valid scenario to consider.